mctp-bench: An utility tool to display throughput.
The recv command waits for incoming packets, notifies the user of
first packet received and then set the polling timeout to the desired number.
Example usage:
	mctp-bench recv

The send command sends packets continously, the size of packet
can be set by user or defaults to 1 byte. User can also set the target eid,
the default value for eid is 8.
Example usage:
	mctp-bench send --size=1000 eid 8

Signed-off-by: Timon Liu <timon@codeconstruct.com.au>